Address slice-5 audit (1 HIGH + 3 MEDIUM in-scope)
Security + architecture lanes flagged; fixes (per operator: fix in-scope, defer
withdraw/#1254 and durable catalog-evidence/#1256):
- HIGH: coordinator served_model_ref validation accepted endpoint material
  embedded in a path segment (ollama:localhost/model, inference:8080/model).
  Now every '/'-segment is checked; the network-location detector treats
  'localhost' as a substring and a numeric host:port as an endpoint, and its
  TLD heuristic requires a pure-alpha last label so model-version tokens
  (Llama-3.2-3B-Instruct-4bit) are not misread as hostnames. Removed the now-dead
  ollama-specific helper; added rejection lock cases.
- MEDIUM: status guidance reported not_earning_yet for non-catalog non-settlement
  states; now returns no_earning_path_in_v0_1 when catalog_model_key is empty
  (SPEC-047-R004 honest v0.1 disclosure).
- MEDIUM: sameModelAdmissionTuple treated catalog_model_key as immutable,
  blocking a legitimate refreshed-evidence re-offer that newly catalog-matches;
  the identity tuple is now CandidateID + ServedModelRef only (catalog binding
  is a dynamic provider claim until verified).
- MEDIUM (layering): extracted JCS canonicalizer to a shared internal/jcs package
  (billing re-exports for its callers) so model_admission.go no longer imports
  internal/billing.

go build/vet + ws admission tests + billing tests green; swift 54 BYOM tests green.
